+++ component/backend/Helper/Select.php
@@ -1,5 +1,11 @@
 <?php
 /**
+ * @package   AkeebaLoginGuard
+ * @copyright Copyright (c)2016-2019 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
+/**
  * @package     Akeeba\LoginGuard\Admin\Helper
  * @subpackage
  *