+++ plugins/loginguard/smsapi/media/js/telinput.js
@@ -1,4 +1,10 @@
 /*
+ * @package   AkeebaLoginGuard
+ * @copyright Copyright (c)2016-2019 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
+/*
  * International Telephone Input v11.0.10
  * https://github.com/jackocnr/intl-tel-input.git
  * Licensed under the MIT license