+++ plugins/loginguard/totp/media/js/qrcode.js
@@ -1,3 +1,9 @@
+/*
+ * @package   AkeebaLoginGuard
+ * @copyright Copyright (c)2016-2019 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
 /*! jquery-qrcode v0.14.0 - https://larsjung.de/jquery-qrcode/ */
 (function (vendor_qrcode) {
     'use strict';