+++ plugins/loginguard/u2f/media/js/u2f-api.js
@@ -1,3 +1,9 @@
+/*
+ * @package   AkeebaLoginGuard
+ * @copyright Copyright (c)2016-2019 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
 //Copyright 2014-2015 Google Inc. All rights reserved.
 
 //Use of this source code is governed by a BSD-style